requirements.txt
fastapi==0.110.1
jupyter==1.0.0
numpy==1.24.4
PyYAML==6.0.1
requests==2.31.0
uvicorn==0.29.0